Removed getpwuid() calls and clarified read_config()
getpwuid() is problematic for portability and cannot be properly linked
statically, whic remains the default. I removed these calls in favor of
querying the environment for $HOME and $USER. This works well with sudo.
I also fixed the value of ConfPath broken by my previous commit to
respect the install prefix and spread out the nested if's as they were
giving me a headache.