Allow spider to run from builddir and make dynamic linking the default
There is really no reason for most users to install spider. I changed the
paths searched for the config. and the default spider.conf path for SSNlin.xml
to the $PWD. When spider is installed with make install sed replaces the path
to SSNlin.xml in the spider.conf installed in @PREFIX@. This is a heavily
requested feature here at the UA. I also made the default linking type dynamic
because the user who is compiling the software doesn't care that it's static
and static libs are rarely easy to come by in the zip,zzip -devel packages on
most distros. When I'll add the --enable-static flag to the spec file
(comming soon) so that we don't have to track deps using rpm.